flask_journal_app_github/requirements.txt
Flask>=3.0.0
python-dateutil>=2.9.0.post0
requests>=2.32.0
